+++ src/jacdac.ts
@@ -1,4 +1,6 @@
+require('./jdconstants')
 require('./jdutils')
 require('./jdpacket')
 require('./jddevice')
+require('./jdbus')
 require('./jdpretty')
